my_project/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
qdrant-client
openai
python-dotenv
pydantic-settings
PyJWT